ProjectLifeCycleStepDefinitions => ProjectPhaseDefinitions
