+++ app/components/settings/project_phase_definitions/index_component.html.erb
@@ -94,7 +94,7 @@ See COPYRIGHT and LICENSE files for more details.
                 }
               ) do
                 render(
-                  Settings::ProjectLifeCycleStepDefinitions::RowComponent.new(
+                  Settings::ProjectPhaseDefinitions::RowComponent.new(
                     definition,
                     first?: definition == definitions.first,
                     last?: definition == definitions.last
