+++ app/components/settings/project_phase_definitions/index_header_component.rb
@@ -29,31 +29,14 @@
 #++
 
 module Settings
-  module ProjectLifeCycleStepDefinitions
-    class RowComponent < ApplicationComponent
-      include OpPrimer::ComponentHelpers
-      include Projects::LifeCycleDefinitionHelper
-
-      alias_method :definition, :model
-
-      options :first?,
-              :last?
-
-      private
-
-      def move_action(menu:, move_to:, label:, icon:)
-        menu.with_item(
-          label:,
-          href: move_admin_settings_project_phase_definition_path(definition, move_to:),
-          form_arguments: {
-            method: :patch
-          },
-          data: {
-            "projects--settings--border-box-filter-target": "hideWhenFiltering"
-          }
-        ) do |item|
-          item.with_leading_visual_icon(icon:)
-        end
+  module ProjectPhaseDefinitions
+    class IndexHeaderComponent < ApplicationComponent
+      def breadcrumbs_items
+        [
+          { href: admin_index_path, text: t("label_administration") },
+          { href: admin_settings_project_custom_fields_path, text: t("label_project_plural") },
+          t("settings.project_phase_definitions.heading")
+        ]
       end
     end
   end
