+++ app/components/settings/project_phase_definitions/row_component.html.erb
