+++ app/views/admin/settings/project_life_cycle_definitions/form.html.erb
@@ -33,6 +33,6 @@ See COPYRIGHT and LICENSE files for more details.
               t("settings.project_phase_definitions.heading"),
               t("settings.project_phase_definitions.#{heading_scope}.heading") %>
 
-<%= render Settings::ProjectLifeCycleStepDefinitions::FormHeaderComponent.new(@definition, heading_scope:) %>
+<%= render Settings::ProjectPhaseDefinitions::FormHeaderComponent.new(@definition, heading_scope:) %>
 
 <%= render Projects::LifeCycleStepDefinitions::FormComponent.new(@definition) %>
